+++ RELEASES.md
@@ -1,3 +1,20 @@
+# v0.3.1
+
+## Major Features
+
+- Bump minimum Go version requirement to v1.17 due to dependencies
+- Preserve preset `cmd.ProcessName` and `cmd.ProcessArgs` if set before invoking
+  `cli.Run()`
+- Add option to disable completion machinery
+- Improve usage display for positional arguments
+
+## Improvements
+
+- Fix typos in test names
+- Update dependencies
+- Add vscode debug configuration
+
+
 # v0.3.0
 
 ## Major Features